Fix chat list showing old messages for unfriended users, move search to navbar
Pending friend requests now always show "Wants to connect" in the chat
sidebar, even when old chat messages exist from a previous friendship.
Moved the search button from AppNav (which disappears when there are no
nav links) to the navbar proper, left of the notification bell.
